requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-multipart>=0.0.12
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
